requirements.txt
requests==2.32.3
fake-useragent==1.5.1
py-cord==2.5.0
python-dotenv==1.0.1